% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_columns.R
\name{align_columns}
\alias{align_columns}
\title{Alignment to qmd table syntax}
\usage{
align_columns(align)
}
\arguments{
\item{align}{A string vector with elements "c" (center), "r" (right), or "l" (left).}
}
\value{
A character vector
}
\description{
Takes an alignment and outputs the qmd table syntax.
Is input into the \code{convert_df_to_qmd_table} function.
}
