% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_theme.R
\name{clean_theme}
\alias{clean_theme}
\title{Clean theme for ggplot}
\usage{
clean_theme()
}
\description{
Creates a custom theme for ggplot, so that all graphs have a coherent look.
}
\examples{
data(donut_data)
library(ggplot2)
ggplot(aes(x = age, y = dist_km), data = donut_data) +
geom_point() +
clean_theme()
}
