% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_qmd_table.R
\name{df_to_qmd_table}
\alias{df_to_qmd_table}
\title{Data frame to qmd table}
\usage{
df_to_qmd_table(df, column_names = NA, align = NA)
}
\arguments{
\item{df}{The data frame that should be converted}

\item{column_names}{The column names that should be printed.
Defaults to \code{names(df)}.}

\item{align}{An optional character vector to define how the columns should be aligned.
Each element can be either "l" (left), "r" (right), or "c" (center).
The position of each element corresponds with the position of each column in \code{column_names}.}
}
\value{
A qmd table.
}
\description{
Converts a data frame or tibble to a qmd table.
This is handy for \code{qmd} Output.
Note: if you want to use this, you need to specify the following in the R chunk:
\code{#| output: asis}
}
\examples{
df_to_qmd_table(iris[1:10,])
}
