% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{donut_data}
\alias{donut_data}
\title{Randomly generated data}
\format{
An object of class \code{data.frame} with 1000 rows and 7 columns.
}
\usage{
donut_data
}
\description{
A data set with which the \code{donut_analysis} functions can be used.
This is in my other package \code{donutsR}.
}
\details{
#' @format ## \code{donut_data}
A data frame with 1000 rows and 7 columns:
\describe{
\item{dist_km}{Distance to nearest point of interest}
\item{wealth_index}{wealth index for observation}
\item{male, age}{gender and age of household head}
\item{id}{ID number for point of interest}
\item{lat, lon}{position of observation}
}
}
\keyword{datasets}
