% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsave_embed.R
\name{ggsave_embed}
\alias{ggsave_embed}
\title{Embed custom font}
\usage{
ggsave_embed(name, width = NA, height = NA)
}
\arguments{
\item{name}{The name of the graph}

\item{width}{The width of the graph. Defaults to NA.}

\item{height}{The height of the graph. Defaults to NA.}
}
\description{
Embed custom fonts in graphs.
}
