% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socials.R
\docType{data}
\name{socials}
\alias{socials}
\title{Sarah's socials}
\format{
An object of class \code{character} of length 1.
}
\usage{
socials
}
\value{
A HTML character vector
}
\description{
Includes my socials for publishing images
}
\examples{
library(ggplot2)
library(ggtext)
ggplot() +
  labs(caption = socials) +
  theme(plot.caption = element_textbox_simple())
}
\keyword{datasets}
