% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_theme.R
\name{clean_theme}
\alias{clean_theme}
\title{Clean theme for ggplot}
\usage{
clean_theme()
}
\description{
Creates a custom theme for ggplot, so that all graphs have a coherent look.
}
\details{
If the fonts are not loaded, please try running the following code.
\code{remove.packages('Rttf2pt1')}
\verb{remotes::install_version('Rttf2pt1', version = '1.3.8')`` }font_import(paths = 'C:/Users/YOUR_NAME/Documents/Fonts')\verb{ `loadfonts(device = 'win')}
`loadfonts()``

If this still does not work, run the following line and then
run the above lines again.
\code{Sys.setenv(R_GSCMD = 'C:/Program Files/gs/gs9.55.0/bin/gswin64c.exe')}"
}
\examples{
data(donut_data)
library(ggplot2)
ggplot(aes(x = age, y = dist_km), data = donut_data) +
geom_point() +
clean_theme()
}
