% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarahsFunctions.R
\docType{package}
\name{sarahsFunctions}
\alias{sarahsFunctions}
\alias{sarahsFunctions-package}
\title{Sarah's custom functions
Often used functions.}
\description{
These are some functions I often use.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/szeller42/sarahsFunctions}
}

}
\author{
Sarah Zeller
}
